Played with the delays and PIN_C pull detection logic.
git-svn-id: svn://localhost/Software/PIC_FillDevice_46K22@330 afd86af0-a409-c442-88d1-3af7309478bb

+++ delay/delay.h
@@ -5,7 +5,7 @@
 extern void	Delay1K(unsigned int cnt);
 
 extern volatile signed int timeout_counter;
-#define set_timeout(a) 	timeout_counter = (int)(((long)(a))/ 10)
+#define set_timeout(a) 	timeout_counter = (((int)(a))/ 10)
 #define is_not_timeout() 	(timeout_counter >= 0)
 
 #define DelayMs(x) 		Delay1K((XTAL_FREQ * ((int)(x))) / 4 )
